# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Information about the API",
        "responses": {
          "200": {
            "description": "The name, version and endpoints of the API",
            "content": {
              "application/json": {
                "example": {
                  "name": "Task API",
                  "version": "1.0",
                  "endpoints": ["/tasks"]
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Check that the server is alive",
        "responses": {
          "200": {
            "description": "The server is running",
            "content": {
              "application/json": {
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Get all tasks, filtered by done or search",
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The list of tasks",
            "content": {
              "application/json": {
                "example": [
                  { "id": 1, "title": "Learn Express", "done": true },
                  { "id": 2, "title": "Build a CRUD API", "done": false },
                  {
                    "id": 3,
                    "title": "Push the project to GitHub",
                    "done": false
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Buy milk"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created task",
            "content": {
              "application/json": {
                "example": { "id": 4, "title": "Buy milk", "done": false }
              }
            }
          },
          "400": {
            "description": "The title is missing or empty",
            "content": {
              "application/json": {
                "example": { "error": "Title is required" }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get one task by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The task with that id",
            "content": {
              "application/json": {
                "example": { "id": 1, "title": "Learn Express", "done": true }
              }
            }
          },
          "404": {
            "description": "No task has that id",
            "content": {
              "application/json": {
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update the title and/or done of a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Buy milk and bread",
                "done": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated task",
            "content": {
              "application/json": {
                "example": {
                  "id": 4,
                  "title": "Buy milk and bread",
                  "done": true
                }
              }
            }
          },
          "400": {
            "description": "The body is empty or invalid",
            "content": {
              "application/json": {
                "example": { "error": "Title or done is required" }
              }
            }
          },
          "404": {
            "description": "No task has that id",
            "content": {
              "application/json": {
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a task by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The task was deleted"
          },
          "404": {
            "description": "No task has that id",
            "content": {
              "application/json": {
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Count the tasks",
        "responses": {
          "200": {
            "description": "How many tasks are done and open",
            "content": {
              "application/json": {
                "example": { "total": 3, "done": 1, "open": 2 }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "Restore the 3 example tasks",
        "responses": {
          "200": {
            "description": "The example tasks",
            "content": {
              "application/json": {
                "example": [
                  { "id": 1, "title": "Learn Express", "done": true },
                  { "id": 2, "title": "Build a CRUD API", "done": false },
                  {
                    "id": 3,
                    "title": "Push the project to GitHub",
                    "done": false
                  }
                ]
              }
            }
          }
        }
      }
    }
  }
}